some stuffs

+++ include/components/transform.h
@@ -1,10 +1,11 @@
 #ifndef TRANSFORM_H
 #define TRANSFORM_H
+#include <stdint.h>
 typedef struct Transform {
-	int x;
-	int y;
-	int rot;
-	int scale;
+	uint8_t x;
+	uint8_t y;
+	uint8_t rot;
+	uint8_t scale;
 } Transform;
 
 #endif
